frontend/requirements.txt
streamlit==1.29.0
python-socketio[client]==5.10.0
requests==2.31.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1